# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sharecrop HTTP API",
    "version": "unversioned",
    "description": "Generated from the route registrations in internal/http/server.go (see internal/openapi). Method, path, operationId, and bearer-auth requirements are derived from the actual mux table and handler bodies. Request/response body schemas are derived from the Go DTO struct each handler actually decodes/writes where that struct could be resolved; an empty schema (`{}`, an unconstrained JSON value) means the handler did not match one of the standard decode/write patterns, not that the body is untyped. See docs/api_reference.md for prose request/response descriptions per route."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/tasks/{task_id}/submissions": {
      "get": {
        "operationId": "listTaskSubmissions",
        "responses": {
          "default": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "submissions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "attachments": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "content_type": {
                                  "type": "string"
                                },
                                "data_url": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                },
                                "size_bytes": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "name",
                                "content_type",
                                "size_bytes",
                                "data_url"
                              ]
                            }
                          },
                          "id": {
                            "type": "string"
                          },
                          "response_json": {
                            "type": "string"
                          },
                          "review_note": {
                            "type": "string"
                          },
                          "sensitive_fields": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "category": {
                                  "type": "string"
                                },
                                "path": {
                                  "type": "string"
                                },
                                "redacted_at": {
                                  "type": "string"
                                },
                                "redaction": {
                                  "type": "string"
                                },
                                "retention": {
                                  "type": "string"
                                },
                                "state": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "path",
                                "category",
                                "retention",
                                "redaction",
                                "state",
                                "redacted_at"
                              ]
                            }
                          },
                          "state": {
                            "type": "string"
                          },
                          "submitter_id": {
                            "type": "string"
                          },
                          "task_id": {
                            "type": "string"
                          },
                          "validation_errors": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "message": {
                                  "type": "string"
                                },
                                "path": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "path",
                                "message"
                              ]
                            }
                          }
                        },
                        "required": [
                          "id",
                          "task_id",
                          "submitter_id",
                          "state",
                          "response_json",
                          "review_note",
                          "attachments",
                          "validation_errors",
                          "sensitive_fields"
                        ]
                      }
                    }
                  },
                  "required": [
                    "submissions"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAuthenticatedSubmission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "content_type": {
                          "type": "string"
                        },
                        "data_url": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "name",
                        "content_type",
                        "data_url"
                      ]
                    }
                  },
                  "response_json": {
                    "type": "string"
                  }
                },
                "required": [
                  "response_json",
                  "attachments"
                ]
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "receipt_token": {
                      "type": "string"
                    },
                    "submission": {
                      "type": "object",
                      "properties": {
                        "attachments": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "content_type": {
                                "type": "string"
                              },
                              "data_url": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "size_bytes": {
                                "type": "integer"
                              }
                            },
                            "required": [
                              "name",
                              "content_type",
                              "size_bytes",
                              "data_url"
                            ]
                          }
                        },
                        "id": {
                          "type": "string"
                        },
                        "response_json": {
                          "type": "string"
                        },
                        "review_note": {
                          "type": "string"
                        },
                        "sensitive_fields": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "category": {
                                "type": "string"
                              },
                              "path": {
                                "type": "string"
                              },
                              "redacted_at": {
                                "type": "string"
                              },
                              "redaction": {
                                "type": "string"
                              },
                              "retention": {
                                "type": "string"
                              },
                              "state": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "category",
                              "retention",
                              "redaction",
                              "state",
                              "redacted_at"
                            ]
                          }
                        },
                        "state": {
                          "type": "string"
                        },
                        "submitter_id": {
                          "type": "string"
                        },
                        "task_id": {
                          "type": "string"
                        },
                        "validation_errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "message": {
                                "type": "string"
                              },
                              "path": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "id",
                        "task_id",
                        "submitter_id",
                        "state",
                        "response_json",
                        "review_note",
                        "attachments",
                        "validation_errors",
                        "sensitive_fields"
                      ]
                    }
                  },
                  "required": [
                    "submission",
                    "receipt_token"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}